refactor: moving building file to one folder
