+++ src/package.json
@@ -1,0 +1,30 @@
+{
+	"name": "@3yourmind/kotti-ui",
+	"version": "0.0.1-beta.4",
+	"description": "Core UI Component",
+	"main": "index.js",
+	"dependencies": {},
+	"author": "3YOURMIND GmbH",
+	"license": "MIT",
+	"repository": {
+		"type": "git",
+		"url": "https://github.com/3YOURMIND/kotti"
+	},
+	"bugs": {
+		"url": "https://github.com/3YOURMIND/kotti/issues"
+	},
+	"homepage": "https://3yourmind.github.io/kotti/inline-edit",
+	"keywords": ["inline edit", "kotti", "ui", "vue", "component"],
+	"contributors": [
+		{
+			"name": "Junyu Pu",
+			"email": "pujunyu@gmail.com",
+			"url": "https://ostoc.github.io/"
+		},
+		{
+			"name": "Kevin Peters",
+			"email": "kevinigeligeligel@gmail.com",
+			"url": "https://www.kevinpeters.net/"
+		}
+	]
+}
